[build] Support gRPC golang bindings in proto_library().
This CL adds a new parameter named 'generate_go_grpc' to the
'proto_library()' template, which is enabled to indicate that a
gRPC go stub, instead of a regular one, should be generated by
the template call.

This will call 'protoc-gen-go-grpc' instead of 'protoc-gen-go'
and will add a suffix of '_grpb.pb.go', instead of '.pb.go' to
the generated source file.

Bug: 72810
Change-Id: I0ae9595ce0ade7611764c1bec294fdcdb8faa9b2