Remove files related to custom_options_test
It turns out that including these tests in the build graph is impossible
because BUILD.gn files in the secondary source location are interpreted
as though they are in the primary source location. This means that paths
are interpreted relative to the primary source location, which doesn't
contain `src/test/custom_options.{cc,proto}`.

Change-Id: Ibb147189981dc8e990a18070950b7162e92fff0e